requirements.txt
fastapi
uvicorn[standard]
pydantic
motor
shapely
geopandas
osmnx
python-dotenv
scikit-learn
pandas
joblib
geopy